Promote getCursorFenceIndices to own file
